+++ plugins/woocommerce-admin/client/wc-api/export/index.js
@@ -1,0 +1,11 @@
+/** @format */
+/**
+ * Internal dependencies
+ */
+import mutations from './mutations';
+import operations from './operations';
+
+export default {
+	mutations,
+	operations,
+};